New variables: IsCivilian Change whether or not this unit is a civilian.
IsHostileCivilian
Change whether or not this unit is an hostile civilian.

CanBeCarried
Change whether or not this unit can be carried when incapacitated

IsAfraidOfFire
Change whether or not this unit will panic when set on fire - only applies to flamethrower

CanTakeCover
Change whether or not can take cover. By default should be true, but certain large units, like mecs and andromedons, etc, don't take cover, so set to false.

WearArmorInBase
Change whether or not this unit should wear their full armor around the base instead of an underlay.